+++ src/common/utils.tsx
@@ -1,0 +1,7 @@
+import { actions as snackActions } from "../store/snack";
+
+export const copyToClipboard = (text: string) => {
+    navigator.clipboard.writeText(text).then(() => {
+      snackActions.showSnack('Text copied to clipboard')
+    })
+  };